+++ crates/shared/ab-proof-of-time/src/lib.rs
@@ -1,10 +1,7 @@
 //! Proof of time implementation.
 
 #![cfg_attr(
-    any(
-        all(not(miri), target_arch = "aarch64"),
-        all(not(miri), target_arch = "x86_64")
-    ),
+    any(all(not(miri), target_arch = "aarch64"), target_arch = "x86_64"),
     feature(portable_simd)
 )]
 #![no_std]